Refactoring interface
Removed code from __init__.py and added to client.py; then updated code in client.py to provide a better interface for class inheritance.
